refactor: 회원 탈퇴 화면 약관 동의 UI 구조 및 위치 개선
* **refactor: `DeletionAgreement` 컴포저블 이동 및 레이아웃 수정**
    * `DeleteAccountNoticeStep` 내부에 위치하던 `DeletionAgreement` 컴포저블을 `DeleteAccountScreen`으로 이동했습니다.
    * 탈퇴 안내(`NOTICE`) 단계에서 하단 `DeleteAccountActionSection` 바로 위에 동의 체크박스가 노출되도록 `DeleteAccountScreen`의 레이아웃 구조를 변경했습니다.
    * 체크박스 및 텍스트 영역에 `noRippleClickable`을 적용하여 클릭 시 상태 변경이 용이하도록 개선했습니다.

* **cleanup: `DeleteAccountNoticeStep` 내 불필요한 코드 정리**
    * 더 이상 사용되지 않는 내부 `DeletionAgreement` 함수 정의를 삭제하고 호출부를 주석 처리 또는 제거하여 코드를 정리했습니다.
