+++ Prezel/feature/setting/impl/src/main/java/com/team/prezel/feature/setting/impl/delete/DeleteAccountScreen.kt
@@ -3,19 +3,29 @@ package com.team.prezel.feature.setting.impl.delete
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalResources
+import androidx.compose.ui.res.stringResource
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.team.prezel.core.designsystem.component.PrezelCheckbox
 import com.team.prezel.core.designsystem.component.feedback.snackbar.showPrezelSnackbar
 import com.team.prezel.core.designsystem.preview.BasicPreview
 import com.team.prezel.core.designsystem.theme.PrezelTheme
 import com.team.prezel.core.ui.state.LocalSnackbarHostState
+import com.team.prezel.core.ui.util.noRippleClickable
 import com.team.prezel.feature.setting.impl.R
 import com.team.prezel.feature.setting.impl.delete.component.DeleteAccountActionSection
 import com.team.prezel.feature.setting.impl.delete.component.DeleteAccountConfirmDialog
@@ -47,6 +57,7 @@ internal fun DeleteAccountScreen(
                     navigateToSplash()
                     snackbarHostState.showPrezelSnackbar(message = resources.getString(R.string.feature_setting_impl_delete_account_withdraw_success))
                 }
+
                 is DeleteAccountUiEffect.ShowMessage -> {
                     snackbarHostState.showPrezelSnackbar(message = resources.getString(effect.message.toMessageRes()))
                 }
@@ -112,12 +123,21 @@ private fun DeleteAccountScreen(
                 .fillMaxWidth(),
         )
 
-        DeleteAccountActionSection(
-            step = uiState.step,
-            enabled = uiState.isPrimaryActionEnabled,
-            onClickNext = onClickNext,
-            onClickWithdraw = onClickWithdraw,
-        )
+        Column(modifier = Modifier.fillMaxWidth()) {
+            if (uiState.step == DeleteAccountStep.NOTICE) {
+                DeletionAgreement(
+                    isChecked = uiState.isNoticeChecked,
+                    onCheckedChange = onToggleNoticeChecked,
+                    modifier = Modifier.padding(horizontal = PrezelTheme.spacing.V20),
+                )
+            }
+            DeleteAccountActionSection(
+                step = uiState.step,
+                enabled = uiState.isPrimaryActionEnabled,
+                onClickNext = onClickNext,
+                onClickWithdraw = onClickWithdraw,
+            )
+        }
     }
 }
 
@@ -143,6 +163,34 @@ private fun DeleteAccountContent(
                 onOtherReasonChanged = onOtherReasonChanged,
             )
         }
+    }
+}
+
+@Composable
+private fun DeletionAgreement(
+    isChecked: Boolean,
+    onCheckedChange: (Boolean) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .noRippleClickable { onCheckedChange(!isChecked) },
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        PrezelCheckbox(
+            checked = isChecked,
+            onCheckedChange = onCheckedChange,
+            extraTouchPadding = PaddingValues(),
+        )
+
+        Spacer(modifier = Modifier.width(PrezelTheme.spacing.V8))
+
+        Text(
+            text = stringResource(R.string.feature_setting_impl_delete_account_notice_check),
+            style = PrezelTheme.typography.body3Regular,
+            color = PrezelTheme.colors.textRegular,
+        )
     }
 }
